Azure/requirements.txt
azure-ai-formrecognizer==3.1.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-core==1.21.1
azure-functions==1.8.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
idna==3.3
isodate==0.6.1
joblib==1.1.0
msrest==0.6.21
nltk==3.6.7
numpy==1.22.1
oauthlib==3.1.1
Pillow==9.0.0
pycparser==2.21
python-dotenv==0.19.2
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.8
Wand==0.6.7
